* updated the gcc ssp patch to it's latest incarnation and adapted 	  the apply logic accordingly (again)
git-svn-id: http://svn.exactcode.de/t2/trunk@4730 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ package/base/gcc/apply-protector.sh
@@ -8,13 +8,12 @@ if [ "$pfile" ] ; then
 
 	if [ -f protector.dif ] ; then
 		patch -p0 < protector.dif
-		mv protector.{c,h} gcc/
 	elif [ -f gcc_*.dif ] ; then
 		patch -p0 < gcc_*.dif
-		# the protector.* files are already extracted into gcc/ ...
 	else
 		abord "Protector patch not found"
 	fi
+	[ -f protector.c ] && mv protector.{c,h} gcc/
 else
 	echo "No stack-protector available for $pkg ..."
 fi
